+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/trygdetid/Trygdetid.tsx
@@ -4,7 +4,7 @@ import { hentTrygdetider, ITrygdetid, opprettTrygdetider } from '~shared/api/try
 import Spinner from '~shared/Spinner'
 import { Alert, BodyShort, Box, Heading, Tabs, VStack } from '@navikt/ds-react'
 import { TrygdeAvtale } from './avtaler/TrygdeAvtale'
-import { IBehandlingStatus, IBehandlingsType, UtlandstilknytningType } from '~shared/types/IDetaljertBehandling'
+import { IBehandlingStatus, IBehandlingsType } from '~shared/types/IDetaljertBehandling'
 import { IBehandlingReducer, oppdaterBehandlingsstatus } from '~store/reducers/BehandlingReducer'
 import { useAppDispatch } from '~store/Store'
 import { isPending } from '~shared/api/apiUtils'
@@ -23,10 +23,8 @@ import { ILand, sorterLand } from '~utils/kodeverk'
 import { ApiErrorAlert } from '~ErrorBoundary'
 import { TrygdetidIAnnenBehandlingMedSammeAvdoede } from '~components/behandling/trygdetid/TrygdetidIAnnenBehandlingMedSammeAvdoede'
 import { FeatureToggle, useFeaturetoggle } from '~useUnleash'
-import { EnigUenigTilbakemelding } from '~shared/tilbakemelding/Tilbakemelding'
-import { ClickEvent } from '~utils/amplitude'
-import { VilkaarsvurderingResultat } from '~shared/api/vilkaarsvurdering'
-import { Revurderingaarsak } from '~shared/types/Revurderingaarsak'
+import { EnigUenigTilbakemelding } from '~shared/tilbakemelding/EnigUenigTilbakemelding'
+import { velgTilbakemeldingClickEventForUtlandsbehandling } from '~utils/amplitude'
 
 interface Props {
   redigerbar: boolean
@@ -39,30 +37,6 @@ const manglerTrygdetid = (trygdetider: ITrygdetid[], avdoede?: Personopplysning[
   const trygdetidIdenter = trygdetider.map((trygdetid) => trygdetid.ident)
   const avdoedIdenter = (avdoede || []).map((avdoed) => avdoed.opplysning.foedselsnummer)
   return !avdoedIdenter.every((ident) => trygdetidIdenter.includes(ident))
-}
-
-function tilbakemeldingForUtlandsbehandling(behandling: IBehandlingReducer): ClickEvent | undefined {
-  // Hvis vi har en førstegangsbehandling utland
-  if (
-    behandling.behandlingType === IBehandlingsType.FØRSTEGANGSBEHANDLING &&
-    !!behandling.utlandstilknytning &&
-    behandling.utlandstilknytning.type !== UtlandstilknytningType.NASJONAL
-  ) {
-    if (behandling.vilkaarsvurdering?.resultat?.utfall === VilkaarsvurderingResultat.IKKE_OPPFYLT) {
-      // Avslag har høyest prioritert
-      return ClickEvent.TILBAKEMELDING_SAKSBEHANDLING_UTLAND_AVSLAG
-    } else if (behandling.erSluttbehandling) {
-      // Hvis ikke avslag -- er dette en sluttbehandling
-      return ClickEvent.TILBAKEMELDING_SAKSBEHANDLING_UTLAND_SLUTTBEHANDLING
-    } else {
-      // Det er en utland førstegangsbehandling
-      return ClickEvent.TILBAKEMELDING_SAKSBEHANDLING_UTLAND_FOERSTEGANGSBEHANDLING
-    }
-  } else if (behandling.revurderingsaarsak === Revurderingaarsak.SLUTTBEHANDLING) {
-    // Revurderinger er sluttbehandlinger hvis det er revurderingsårsak sluttbehandling utland
-    return ClickEvent.TILBAKEMELDING_SAKSBEHANDLING_UTLAND_SLUTTBEHANDLING
-  }
-  return undefined
 }
 
 export const Trygdetid = ({ redigerbar, behandling, vedtaksresultat, virkningstidspunktEtterNyRegelDato }: Props) => {
@@ -160,7 +134,7 @@ export const Trygdetid = ({ redigerbar, behandling, vedtaksresultat, virkningsti
     )
   }
 
-  const tilbakemeldingUtland = tilbakemeldingForUtlandsbehandling(behandling)
+  const tilbakemeldingClickEventForUtlandsbehandling = velgTilbakemeldingClickEventForUtlandsbehandling(behandling)
 
   return (
     <Box paddingInline="16" maxWidth="69rem">
@@ -265,11 +239,11 @@ export const Trygdetid = ({ redigerbar, behandling, vedtaksresultat, virkningsti
               : 'Finner ikke trygdetid - ID mangler'}
           </ApiErrorAlert>
         )}
-        {tilbakemeldingUtland && (
+        {tilbakemeldingClickEventForUtlandsbehandling && (
           <Box paddingBlock="0 16">
             <EnigUenigTilbakemelding
               spoersmaal="Jeg synes det er lett å behandle utlandssaker i Gjenny"
-              clickEvent={tilbakemeldingUtland}
+              clickEvent={tilbakemeldingClickEventForUtlandsbehandling}
               behandlingId={behandling.id}
             />
           </Box>
